split data from chart
